--- MyApplication2/.idea/.name
@@ -1,1 +1,0 @@
-WeAllCode1
