--- MyApplication2/app/src/androidTest/java/com/wesleycoombergmail/weallcode1/ApplicationTest.java
@@ -1,13 +1,0 @@
-package com.wesleycoombergmail.weallcode1;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
-    }
-}
