+++ WeAllCode/settings.gradle
